better handling of handshake.
we checked for venti-02-.  02 is the version.  but actually the
version sent is a list of colon-separated supported versions.  p9p's
vac/venti tools send venti-04:02-.  we can now talk to such clients
again.

while here, it seems the problem on openbsd has gone away, so remove
warning in manual page.